DRA integration: split source code
dra.go grew organically over time. Some tests where added there, others in
separate files. Separate files is better because it's easier to find tests and
because it avoids merge conflicts. To make the approach consistent, all code
that isn't related to how tests are run gets moved out of dra.go.
